Separated commands into files
