+++ commands/github-issue-number.js
@@ -1,6 +1,6 @@
 module.exports = {
-	name: "github-info",
-	description: 'Get GitHub Personal Access Token',
+	name: "github-issue-number",
+	description: 'GitHub bot command 2',
 	execute(command, message, args) {
 		// Intialise GitHub API
     const github = require("octonode");
@@ -21,15 +21,15 @@ module.exports = {
     const ghuser = client.user("Inoxia25");
     const ghrepo = client.repo("MLH-Fellowship/pod-3.1.3-team-4");
 
-    // -github-info: Get contents of personal token
-    if (command === "github-info") {
-      if (!args.length) {
-        return message.reply("you didn't provide a GitHub Personal Token.");
-      } else {
-      githubToken = args[0];
-      client = github.client(githubToken);
-      return message.reply("GitHub auth was successful. Use -github-issue-number with the number of the issue you'd like to comment on.");
-      }
+    // -github-issue-number: Get issue/PR number from user
+    if (command === "github-issue-number") {
+        issue = parseInt(args[0]);
+  
+        if (isNaN(issue)) {
+          return message.reply("sorry, that isn't a valid issue number.").catch(error => console.log(error));
+        } else {
+          return message.reply("to post a comment on issue " + issue + " , use -github-post-comment followed by your message.");
+        }
+      } 
     }
-}
 }
